+++ vscode/src/autoedits/renderer/image-gen/highlight/constants.ts
@@ -1,0 +1,6 @@
+import type { SYNTAX_HIGHLIGHT_THEME } from './types'
+
+export const DEFAULT_HIGHLIGHT_COLORS: Record<SYNTAX_HIGHLIGHT_THEME, string> = {
+    dark: '#ffffff',
+    light: '#000000',
+}
