+++ vscode/src/autoedits/renderer/image-gen/highlight/types.ts
@@ -1,0 +1,1 @@
+export type SYNTAX_HIGHLIGHT_THEME = 'light' | 'dark'
